Remove BlockValidator interface.
