--- ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/statetransition/blockvalidator/NoOpBlockValidator.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2020 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.spec.logic.common.statetransition.blockvalidator;
-
-import tech.pegasys.teku.spec.cache.IndexedAttestationCache;
-import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
-import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
-
-class NoOpBlockValidator implements BlockValidator {
-
-  @Override
-  public BlockValidationResult validatePreState(
-      BeaconState preState,
-      SignedBeaconBlock block,
-      IndexedAttestationCache indexedAttestationCache) {
-    return BlockValidationResult.SUCCESSFUL;
-  }
-
-  @Override
-  public BlockValidationResult validatePostState(BeaconState postState, SignedBeaconBlock block) {
-    return BlockValidationResult.SUCCESSFUL;
-  }
-}
